+++ src/Vehicle/Vehicle.java
@@ -4,9 +4,9 @@ public abstract class Vehicle {
     String plateNumber;
     String color;
     String brand;
-    int baseRate;
+    double baseRate;
 
-    Vehicle(String plateNumber, String color, String brand, int baseRate) {
+    Vehicle(String plateNumber, String color, String brand, double baseRate) {
         this.plateNumber = plateNumber;
         this.color = color;
         this.brand = brand;
@@ -25,14 +25,15 @@ public abstract class Vehicle {
         return brand;
     }
 
-    public int getBaseRate() {
+    public double getBaseRate() {
         return baseRate;
     }
+
     public String getName() {
         return brand;
     }
 
     public abstract String getType();
-    public abstract int calculateRate(int distance);
+    public abstract double calculateRate(double distance);
     public abstract void showVehicle();
 }
